pyproject.toml
[tool.poetry]
name = "not-llama-fs"
version = "0.1.0"
description = ""
authors = ["drforse <george.lifeslice@gmail.com>"]
readme = "README.md"
packages = [{include = "not_llama_fs"}]

[tool.poetry.dependencies]
python = "^3.10"
ollama = "^0.2.0"
python-magic = "^0.4.27"
python-magic-bin = "^0.4.14"
asciitree = "^0.3.3"
groq = "^0.8.0"
openai = "^1.30.5"
anthropic = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
